Move database config folder to server folder
