+++ server/models/index.js
@@ -6,7 +6,7 @@ require('dotenv').config();
 
 const basename = path.basename(module.filename);
 const env = process.env.NODE_ENV || 'development';
-const config = require('../../config/config.json')[env];
+const config = require('../config/config.json')[env];
 
 const db = {};
